removing attribute retrieval by string for bool, int and float
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@3877 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/xml/WeightsHandler.cpp
@@ -131,8 +131,8 @@ void WeightsHandler::myStartElement(int element, const std::string &/*name*/,
     switch (element) {
     case SUMO_TAG_INTERVAL:
         try {
-            myCurrentTimeBeg = getLong(attrs, SUMO_ATTR_BEGIN);
-            myCurrentTimeEnd = getLong(attrs, SUMO_ATTR_END);
+            myCurrentTimeBeg = getInt(attrs, SUMO_ATTR_BEGIN);
+            myCurrentTimeEnd = getInt(attrs, SUMO_ATTR_END);
         } catch (...) {
             MsgHandler::getErrorInstance()->inform("Timestep value is not numeric.");
         }
